07_blog_generation_agenticAI_using_crewAI/requirements.txt
crewai
crewai-tools # for third party tools
python-dotenv
streamlit